Mid term changes

+++ src/main/java/at/fhtw/swen3/persistence/repositories/ParcelRepository.java
@@ -4,8 +4,8 @@ import at.fhtw.swen3.persistence.entities.ParcelEntity;
 import at.fhtw.swen3.persistence.entities.RecipientEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ParcelRepository extends JpaRepository<ParcelEntity, Long> {
-    ParcelEntity findById(long id);
+public interface ParcelRepository extends JpaRepository<ParcelEntity, Integer> {
+    ParcelEntity findById(int id);
     ParcelEntity findByTrackingId(String trackingId);
     ParcelEntity findBySender(RecipientEntity sender);
     ParcelEntity findByRecipient(RecipientEntity recipient);
